requirements.txt

asgiref==3.8.1
boto3==1.34.146
botocore==1.34.146
Django==5.0.7
jmespath==1.0.1
numpy==2.0.1
pandas==2.2.2
pillow==10.4.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
s3transfer==0.10.2
six==1.16.0
sqlparse==0.5.1
tzdata==2024.1
urllib3==2.2.2
gunicorn==20.0.0
dj-database-url==2.2.0
psycopg2-binary==2.9.9
djangorestframework==3.15.2
